Un-depend on Tk
AnyEvent supports numerous event loop implementations, including Tk. It
does not require *all* of them to be present. While depending on most is
lightweight and convenient, Tk pulls in a significant amount of
graphical support, greatly increasing the size of the perl-basic bundle.

Removing the RPM dependency on Tk here means that all the other event
models will still work as before, but if a user wants to use the Tk
model, they'll need to add another bundle (typically perl-extras) to get
it. However, if someone wants to use a Tk event loop, they're likely
already working with a Tk program, and probably would have added the
bundle anyway.